feat: migrate quantities and stock to BigDecimal

+++ server/src/main/java/com/store/api/dto/ItemRequest.java
@@ -10,8 +10,8 @@ public class ItemRequest {
     private BigDecimal price;
     private BigDecimal cost;
     private String imageUrl;
-    private Integer stock;
-    private Integer minStock;
+    private BigDecimal stock;
+    private BigDecimal minStock;
     private Long categoryId;
     private Long supplierId;
     private Boolean visible;
